Rename `BondedECDSAKeepFilterer`
